chore: create separate test cases for register view
