+++ packages/client/src/v2-events/features/events/actions/validate/Pages.stories.tsx
@@ -11,11 +11,7 @@
 import type { Meta, StoryObj } from '@storybook/react'
 import { createTRPCMsw, httpLink } from '@vafanassieff/msw-trpc'
 import superjson from 'superjson'
-import {
-  ActionType,
-  generateEventDocument,
-  tennisClubMembershipEvent
-} from '@opencrvs/commons/client'
+import { tennisClubMembershipEvent } from '@opencrvs/commons/client'
 import { ROUTES, routesConfig } from '@client/v2-events/routes'
 import { tennisClubMembershipEventDocument } from '@client/v2-events/features/events/fixtures'
 import { AppRouter } from '@client/v2-events/trpc'
